refactor: prepare backend for vercel serverless
