+++ api/index.ts
@@ -1,0 +1,2 @@
+import app from "../server/app";
+export default app;
