+++ server/db.ts
@@ -1,10 +1,7 @@
-import { Pool, neonConfig } from '@neondatabase/serverless';
-import { drizzle } from 'drizzle-orm/neon-serverless';
-import ws from "ws";
+import mysql from "mysql2/promise";
+import { drizzle } from "drizzle-orm/mysql2";
 import 'dotenv/config';
 import * as schema from "@shared/schema";
-
-neonConfig.webSocketConstructor = ws;
 
 if (!process.env.DATABASE_URL) {
   throw new Error(
@@ -12,5 +9,7 @@ if (!process.env.DATABASE_URL) {
   );
 }
 
-export const pool = new Pool({ connectionString: process.env.DATABASE_URL });
-export const db = drizzle({ client: pool, schema });
+const pool = mysql.createPool(process.env.DATABASE_URL);
+// Cast to any to allow using existing Postgres-oriented schema
+export const db: any = drizzle(pool as any, { schema: schema as any });
+export { pool };
